+++ app/src/main/java/curso/alberto/sunriseweatherapp/data-retrofit/Previccion_Hora.kt
@@ -126,7 +126,7 @@ data class Ice (
     @SerializedName("UnitType" ) var UnitType: Int? = null
 )
 data class Evapotranspiration (
-    @SerializedName("Value") var Value: Int? = null,
+    @SerializedName("Value") var Value: Double? = null,
     @SerializedName("Unit" ) var Unit: String? = null,
     @SerializedName("UnitType" ) var UnitType: Int? = null
 )
